requirements.txt
requests>=2.28.0
python-frontmatter>=1.0.0
PyYAML>=6.0
flask>=2.3.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0
APScheduler>=3.10.0
